image-caption-service/pyproject.toml
[project]
name = "image-caption-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.1",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]
